Generate 40-Year Average Monthly Temperature Plot

The function generate_40_year_average_temperature_plot() generates a plot showing the average temperature of the specified weather station for each month over 40-year intervals. This allows for easy comparison of average temperatures across different months for each 40-year period available in the dataset.

Usage

generate_40_year_average_temperature_plot(id)

Arguments

id

Character string representing the ID of the weather station for which to generate the plot. This should match the station IDs in the weatherdata dataset.

Value

A ggplot object displaying the average monthly temperature for the specified weather station across 40-year intervals.

Examples

generate_40_year_average_temperature_plot("NH0472")
#> Warning: data set ‘weatherdata’ not found
#> Error in generate_40_year_average_temperature_plot("NH0472"): object 'weatherdata' not found